Update rendering components
