+++ src/solace_ai_connector_web/frontend/app/components/PreviewFileContent/Renderers/types.ts
@@ -1,0 +1,4 @@
+export interface BaseRendererProps {
+    content: string;
+    width: number;
+  }
